requirements.txt
Flask
psycopg2-binary
python-dotenv==3.0.0
gunicorn
Werkzeug>=3.0.0
